config_system: Add new config system using structs
The "old" config system is quite rigid in terms of user
parameterization. To avoid adding more and more generics to the
parameter port I've packaged the parameters to use a single struct
`ariane_cfg_t` which can then be passed to all submodules in need of the
parameter value.

We can slowly transition to this new system. One drawback is the missing
capability to calculate parameters based on other parameters. Functions
would be needed to do so. Or, imho, the better approach would be a
Python script which generates the appropriate parameters.